rename: Api2 -> Api로 임시 네이밍 변경 사항 수정
